% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close.r
\name{compute_close}
\alias{compute_close}
\title{Conditional Location Scale Empirical Bayes (CLOSE)}
\usage{
compute_close(
  estimates,
  standard_errors,
  conditional_mean = NULL,
  conditional_std = NULL,
  truncate = 1e-08,
  grid = NULL,
  ...
)
}
\arguments{
\item{estimates}{A vector of noisy estimates for some underlying parameters. It is assumed
that these estimates are Gaussian with standard deviation equal to \code{standard_errors}
and means equal to the underlying parameters.}

\item{standard_errors}{A vector of standard errors for the estimates.}

\item{conditional_mean}{A vector of estimated conditional mean.
\code{E[estimates | standard_error]} for the parameters. If \code{NULL}, this will be estimated by
\code{local_linear_regression_conditional_moments}.}

\item{conditional_std}{A vector of estimated conditional standard deviation
for the parameters: \code{Sqrt(Var(estimates | standard_errors) - standard_errors^2)}.
If \code{NULL}, this will be estimated by \code{local_linear_regression_conditional_moments}.}

\item{truncate}{A small number. If the estimated conditional standard deviation is
less than truncate, the observations are ignored.}

\item{grid}{A grid of values to use for the NPMLE estimate of the conditional distribution
of the parameters. If \code{NULL}, a default grid is used. The default grid is

\if{html}{\out{<div class="sourceCode">}}\preformatted{grid <- seq(-6, 6, length.out=400)
minx <- min(-7, min(transformed_estimates[good_obs]))
max <- max(7, max(transformed_estimates[good_obs]))
grid <- c(seq(minx, -6, length.out=100)[1:99], grid, seq(6, max, length.out=100)[2:100])
}\if{html}{\out{</div>}}}

\item{...}{Additional arguments to be passed to \code{local_linear_regression_conditional_moments}.
In particular, if one wishes to estimate the conditional moment based on some transformation of
\code{standard_errors}, one should pass in, e.g., \code{morphed_standard_errors = log(standard_errors)}
here.}
}
\value{
A list with the following elements:
\itemize{
\item{\code{close_gauss}} The posterior means according to CLOSE-GAUSS
\item{\code{close_npmle}} The posterior means according to CLOSE-NPMLE
\item{\code{conditional_mean}}  The estimated conditional mean of the parameters
\item{\code{conditional_std}}  The estimated conditional standard deviation of the parameters
\item{\code{npmle}}  The metadata on NPMLE returned by \code{REBayes::GLmix}
}
}
\description{
This function takes in estimates and standard errors and computes
CLOSE-GAUSS and CLOSE-NPMLE empirical Bayes posterior means.
}
\examples{
n <- 10000
standard_errors <- 0.1 + runif(n)
thetas <- standard_errors + 0.3 * rnorm(n)
estimates <- thetas + standard_errors * rnorm(n)
close_results <- compute_close(estimates, standard_errors)

}
