% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRP.R
\name{CRP}
\alias{CRP}
\title{CLOSE-R main functions, estimate copy number based on CRP
calls all the sub functions inside of subFunc.R}
\usage{
CRP(Input, sampleName = "closer")
}
\arguments{
\item{Input}{segments with LAF and LRR. See manual for more detail}

\item{sampleName}{output files prefix. "closer" is not specified}

\item{codePath}{directory for subFunc.R}
}
\description{
CLOSE-R main functions, estimate copy number based on CRP
calls all the sub functions inside of subFunc.R
}

