% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAP.R
\name{dmode}
\alias{dmode}
\title{Function to find the mode of a distribution}
\source{
http://stackoverflow.com/questions/16255622/peak-of-the-kernel-density-estimation
}
\usage{
dmode(x)
}
\arguments{
\item{x}{}
}
\description{
Function to find the mode of a distribution
}

