% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prior_moments.r
\name{fast_bandwidth_select}
\alias{fast_bandwidth_select}
\title{Use a subset of the observations to select bandwidth}
\usage{
fast_bandwidth_select(x, y, frac = 0.2, ...)
}
\arguments{
\item{x}{A vector of the x-values.}

\item{y}{A vector of the y-values.}

\item{frac}{A number between 0 and 1 denoting the fraction of observations to use.}

\item{...}{Additional arguments to be passed to \code{llr_bwselect_wrapper}.}
}
\value{
A list with the following elements:
\itemize{
\item{\code{h}} The selected bandwidth.
\item{\code{kernel}} The kernel used.
}
}
\description{
This function is a wrapper for \code{llr_bwselect_wrapper} that
first picks \code{frac} of the observations at random and
pass to \code{llr_bwselect_wrapper}.
}
\examples{
x <- runif(10000)
y <- x + rnorm(10000)
bw_result <- fast_bandwidth_select(x, y, frac = 0.2)
}
