% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findpeaks.R
\name{findPeaks}
\alias{findPeaks}
\title{find the density peaks of minor allele copy number ratio (minCNR)}
\usage{
findPeaks(y)
}
\arguments{
\item{y}{minCNR for all segments; the first element of the list returned by getASCN()}
}
\value{
a list with two elements (fit_mat , peak_mat)  
# 1) fit_mat: Minor_Allelic_CN, density for each segments
2) peak_mat: peak_value, peak_y for each identified peak
}
\description{
find the density peaks of minor allele copy number ratio (minCNR)
}

