% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnstatus.R
\name{getASCN}
\alias{getASCN}
\title{calualte minCNR (minor copy number ratio), majCNR (major copy number ratio)}
\usage{
getASCN(LAF, R, scale = 1)
}
\arguments{
\item{LAF}{vector of LAF values}

\item{R}{ratio of read depth (tumor/normal) for each segments: usualy 2^LRR}

\item{scale}{scale factor for library size}
}
\value{
list with 2 elements (minCNR and majCNR for each segment)
}
\description{
calualte minCNR (minor copy number ratio), majCNR (major copy number ratio)
}

