% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnstatus.R
\name{getCentroid}
\alias{getCentroid}
\title{find centroid of the CNR clsuters}
\usage{
getCentroid(segs.cluster, Ncluster)
}
\arguments{
\item{segs.cluster}{}

\item{Ncluster}{number of clusters; third element of the list returned from runDP()}
}
\value{
centroid of each cluster
}
\description{
find centroid of the CNR clsuters
}

