% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnstatus.R
\name{getCNstatus}
\alias{getCNstatus}
\title{estimate the copy number status: normal, gainm loss, LOH, ambigous}
\usage{
getCNstatus(centroid, CNR.mat, segs, segs.cluster, Ncluster)
}
\arguments{
\item{centroid}{}

\item{CNR.mat}{}

\item{segs}{}

\item{segs.cluster}{}

\item{Ncluster}{}
}
\description{
estimate the copy number status: normal, gainm loss, LOH, ambigous
}

