% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prior_moments.r
\name{llr_bwselect_wrapper}
\alias{llr_bwselect_wrapper}
\title{Wrapper for \code{nprobust::lpbwselect}}
\usage{
llr_bwselect_wrapper(
  x,
  y,
  n_real = NULL,
  kernel = "epa",
  bwselect = "imse-dpi",
  ...
)
}
\arguments{
\item{x}{A vector of the x-values.}

\item{y}{A vector of the y-values.}

\item{n_real}{An integer equal to the sample size. Optional and defaults to \code{length(x)}.}

\item{kernel}{A string denoting a kernel type. See \code{nprobust::lpbwselect}.
Optional and defaults to \code{epa} for the Epanechnikov kernel.}

\item{bwselect}{A string denoting a bandwidth selection method.
See \code{nprobust::lpbwselect}. Defaults to \code{imse-dpi}.}

\item{...}{Additional arguments to be passed to \code{nprobust::lpbwselect}.}
}
\value{
A list with the following elements:
\itemize{
\item{\code{h}} The selected bandwidth.
\item{\code{kernel}} The kernel used.
}
}
\description{
This function wraps \code{nprobust::lpbwselect} and returns the selected
optimal bandwidth as a function of data and kernel.
Only supports local linear regression. Supports using a
subset of the data to estimate the optimal bandwidth by \code{n_real}.
That is, the bandwidth is equal to \code{h * (n_real / length(x))^-0.2}.
If \code{length(x) < 40}, then the bandwidth
\code{h <- 1.06 * sd(x) * (length(x) + 3)^-0.2}
is returned, and a warning raised.
}
\examples{
x <- runif(100)
y <- x + rnorm(100)
bw_result <- llr_bwselect_wrapper(x, y)
}
