% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prior_moments.r
\name{local_linear_regression}
\alias{local_linear_regression}
\title{Local linear regression}
\usage{
local_linear_regression(x, y, kernel = "epa", bwselect = "imse-dpi", ...)
}
\arguments{
\item{x}{A numeric vector of x-values}

\item{y}{A numeric vector of y-values}

\item{kernel}{The kernel to use. See \code{nprobust::lpbwselect} for details. Defaults to \code{epa}.}

\item{bwselect}{The bandwidth selection method to use.
See \code{nprobust::lpbwselect} for details. Defaults to \code{imse-dpi}.}

\item{...}{Additional arguments to pass to \code{fast_bandwidth_select}.}
}
\value{
A list with the following elements:
\itemize{
\item{\code{f_hat}} A vector that contains the estimated value
of the conditional expectation at each x-value.
\item{\code{eff_sample_size}} A number equal to the average effective
sample size averaged over evaluation points. The effective sample
size is defined as the inverse sum of squares of the weights used
at the evaluation point. That is, if
\verb{f_hat(x) = sum_i w_i(x) y_i},
then \verb{eff_sample_size_i = 1 / sum_i w_i(x)^2}.}
}
\description{
This function implements local linear regression .
}
\examples{
x <- runif(10000)
y <- x + rnorm(10000)
llr_result <- local_linear_regression(x, y)
}
