% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prior_moments.r
\name{local_linear_regression_conditional_moments}
\alias{local_linear_regression_conditional_moments}
\title{Estimation of conditional moments of \code{theta | sigma} using local linear regression}
\usage{
local_linear_regression_conditional_moments(
  estimates,
  standard_errors,
  morphed_standard_errors = NULL,
  kernel = "epa",
  bwselect = "imse-dpi",
  fast = TRUE,
  variance_fit_type = "squared_residual",
  truncate = TRUE,
  ...
)
}
\arguments{
\item{estimates}{A numeric vector of noisy estimates (\code{Y}).}

\item{standard_errors}{A numeric vector of standard errors (\code{sigma}).}

\item{morphed_standard_errors}{A numeric vector of transformed
standard errors (e.g. \code{log(sigma)}) to estimate the conditional moments with.
Defaults to \code{standard_errors}.}

\item{kernel}{The kernel to use. See \code{nprobust::lpbwselect} for details. Defaults to \code{epa}.}

\item{bwselect}{The bandwidth selection method to use. Defaults to \code{imse-dpi}.}

\item{fast}{Whether to use \code{local_linear_regression_fast} or \code{local_linear_regression}.}

\item{variance_fit_type}{The type of variance estimator to use. The choices are
\code{squared_residual} (the default), \code{fit_conditional}, and \code{difference_of_squares}.
\itemize{
\item{\code{squared_residual}} The variance estimator is an analogue estimator of
\code{Var(theta | sigma) = E[(Y - E[Y | sigma])^2 | sigma] - sigma^2}.
\item{\code{fit_conditional}} The variance estimator is an analogue estimator of
\code{Var(theta | sigma) = E[(Y - E[Y | sigma])^2 - sigma^2 | sigma]}.
\item{\code{difference_of_squares}} The variance estimator is an analogue estimator of
\code{Var(theta | sigma) = E[Y^2 | sigma^2] - E[E[Y | sigma]^2 | sigma] - sigma^2}.
}}

\item{truncate}{Whether to truncate based on
the Kubokawa, Robert, and Saleh (Canadian Journal of Statistics, 1993) heuristic.
Defaults to \code{TRUE}.
If TRUE, then the conditional variance is truncated below at
\code{2/(eff_sample_size + 2) * E[(Y - E[Y | sigma])^2 | sigma]},
where \code{E[(Y - E[Y | sigma])^2 | sigma]} is an estimated analogue
Otherwise, the conditional variance is truncated below at zero.}

\item{...}{Additional arguments to pass to
\code{local_linear_regression_fast} or \code{local_linear_regression}.}
}
\value{
A list with the following elements:
\itemize{
\item{\code{conditional_mean}} A numeric vector of the estimated conditional mean
\item{\code{conditional_std}} A numeric vector of the estimated conditional standard deviation
\item{\code{eff_sample_size}} A number equal to the average effective
sample size averaged over evaluation points. See \code{local_linear_regression}.
}
}
\description{
This function estimates the conditional mean and variance of \code{theta | sigma}
using local linear regression.
}
\examples{
n <- 10000
standard_errors <- 0.1 + runif(n)
thetas <- standard_errors + 0.3 * rnorm(n)
estimates <- thetas + standard_errors * rnorm(n)
conditional_moments <- local_linear_regression_conditional_moments(
    estimates = estimates, standard_errors = standard_errors
)
}
