% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_prior_moments.r
\name{local_linear_smoothing}
\alias{local_linear_smoothing}
\title{Local linear regression}
\usage{
local_linear_smoothing(x, y, h, evals = NULL, kernel = "epa")
}
\arguments{
\item{x}{A vector of the x-values.}

\item{y}{A vector of the y-values.}

\item{h}{A number denoting the bandwidth.}

\item{evals}{A vector of evaluation points. Optional and defaults to x.}

\item{kernel}{A string denoting a kernel type. See \code{nprobust::lpbwselect}.
Defaults to "epa" for the Epanechnikov kernel.}
}
\value{
A list with the following elements:
\itemize{
\item{\code{f_hat}} A vector that contains the estimated value
of the conditional expectation at each x-value.
\item{\code{eff_sample_size}} A number equal to the average effective
sample size averaged over evaluation points. The effective sample
size is defined as the inverse sum of squares of the weights used
at the evaluation point. That is, if \verb{f_hat(x) = sum_i w_i(x) y_i},
then \verb{eff_sample_size_i = 1 / sum_i w_i(x)^2}.
}
}
\description{
This function implements local linear regression with a given
bandwidth and evaluation points. The result is linearly interpolated
from the evaluation points to all x-values.
}
\examples{
x <- runif(100)
y <- x + rnorm(100)
bw_result <- llr_bwselect_wrapper(x, y)
llr_result <- local_linear_smoothing(x, y, bw_result$h)
}
