% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_dp.R
\name{runDP}
\alias{runDP}
\title{cluster segments based on LAF_median and LRR_mean using dirichlet process}
\source{
http://statistical-research.com/dirichlet-process-infinite-mixture-models-and-clustering/
}
\usage{
runDP(x, y, disp.param = 0.45, max.iter = 100, tolerance = 0.001)
}
\arguments{
\item{x}{x variable (LAF for each segment)}

\item{y}{y variable (LRR for each segment)}

\item{disp.param}{}

\item{max.iter}{}

\item{tolerance}{}
}
\description{
cluster segments based on LAF_median and LRR_mean using dirichlet process
}

